ARM: Minimal support for the hard lockup dectector
Currently there is no support for the hard lockup detector (a.k.a. the
NMI watchdog) for ARM.

This patch provides the absolute bare minimum needed to allow the
watchdog to operate and does not tackle a few complications.

In particular boot ordering issues prevent the watchdog from starting
without assistance from userspace (the watchdog is initialized before
the PMU). Likewise there is no logic to automatically disable the
watchdog on systems with no support for FIQ (the watchdog is safe to run
from IRQ but it becomes ineffective).

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>