irq: gic: Add nofiq option
Provide a nofiq option to allow sub-systems for which the use of FIQ is
options (like perf) to be switched between IRQ and FIQ mode.

This is great for live demos and can also be used to test some of the
error paths in the irq subsystem.

Note: However useful this patch is it is a hack and will not be
      upstreamed.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>